minor dependency cleanup, split eager logger configuration to set bean-utils log level
