+++ warp-core-gatling/locking/warp-core-gatling_2.13.lockfile
@@ -59,10 +59,11 @@ com.typesafe.slick:slick-hikaricp_2.13:3.4.1=compileClasspath,default,runtimeCla
 com.typesafe.slick:slick_2.13:3.4.1=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
 com.typesafe:config:1.4.0=scalastyle
 com.typesafe:config:1.4.2=compileClasspath,testCompileClasspath
-com.typesafe:config:1.4.3=default,runtimeClasspath,testRuntimeClasspath
+com.typesafe:config:1.4.3=default
+com.typesafe:config:1.4.4=runtimeClasspath,testRuntimeClasspath
 com.zaxxer:HikariCP:4.0.3=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
-commons-beanutils:commons-beanutils:1.9.3=checkstyle
-commons-beanutils:commons-beanutils:1.9.4=default,runtimeClasspath,testRuntimeClasspath
+commons-beanutils:commons-beanutils:1.9.3=checkstyle,runtimeClasspath,testRuntimeClasspath
+commons-beanutils:commons-beanutils:1.9.4=default
 commons-cli:commons-cli:1.4=checkstyle
 commons-collections:commons-collections:3.2.2=checkstyle,default,runtimeClasspath,testRuntimeClasspath
 commons-io:commons-io:2.16.0=default,runtimeClasspath,testRuntimeClasspath
