+++ warp-core/locking/warp-core_2.12.lockfile
@@ -38,12 +38,13 @@ com.swoval:file-tree-views:2.1.12=zinc
 com.typesafe.slick:slick-hikaricp_2.12:3.4.1=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
 com.typesafe.slick:slick_2.12:3.4.1=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
 com.typesafe:config:1.4.0=scalastyle
-com.typesafe:config:1.4.3=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
+com.typesafe:config:1.4.3=default
+com.typesafe:config:1.4.4=compileClasspath,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
 com.zaxxer:HikariCP:4.0.3=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
-commons-beanutils:commons-beanutils:1.9.3=checkstyle
-commons-beanutils:commons-beanutils:1.9.4=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
+commons-beanutils:commons-beanutils:1.9.3=checkstyle,runtimeClasspath,testRuntimeClasspath
+commons-beanutils:commons-beanutils:1.9.4=default
 commons-cli:commons-cli:1.4=checkstyle
-commons-collections:commons-collections:3.2.2=checkstyle,compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
+commons-collections:commons-collections:3.2.2=checkstyle,default,runtimeClasspath,testRuntimeClasspath
 commons-io:commons-io:2.16.0=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
 commons-logging:commons-logging:1.3.2=compileClasspath,default,runtimeClasspath,testCompileClasspath,testRuntimeClasspath
 io.github.java-diff-utils:java-diff-utils:4.12=zinc
