+++ warp-core/src/main/scala/com/workday/warp/config/WarpPropertyManager.scala
@@ -3,11 +3,10 @@ package com.workday.warp.config
 import java.io.File
 import java.util.Properties
 import com.workday.warp.inject.WarpGuicer
-import com.workday.warp.logger.WarpLogUtils
+import com.workday.warp.logger.{LoggerInit, WarpLogUtils}
 import org.apache.commons.configuration2.PropertiesConfiguration
 import org.apache.commons.configuration2.builder.fluent.Configurations
 import org.apache.commons.configuration2.ex.ConfigurationException
-import org.slf4j.{Logger, LoggerFactory}
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable
@@ -23,7 +22,7 @@ import scala.util.{Failure, Success, Try}
   * Created by tomas.mccandless on 10/21/15.
   * Based on a class created by michael.ottati on 3/29/13
   */
-object WarpPropertyManager {
+object WarpPropertyManager extends LoggerInit {
 
   /**
     * Properties containing this prefix will be passed through as system properties with this
@@ -33,9 +32,6 @@ object WarpPropertyManager {
   private val SYSTEM_PROPERTY_PREFIX: String = "systemProp."
   private val WARP_PROPERTIES: String = "warp.properties"
   private val WARP_CONFIG_DIRECTORY_PROPERTY: String = "wd.warp.config.directory"
-
-  @transient
-  protected lazy val logger: Logger = LoggerFactory.getLogger(getClass.getName)
 
   // immutable copy of jvm system properties
   val systemProps: Map[String, String] = System.getProperties.asScala.toMap
